requirements.txt
pandas
matplotlib
seaborn
jupyter
google-generativeai
python-dotenv
tabulate
qdrant-client
qdrant-client
PyYAML
